# authz: cache duration metric for subrepo perms