@keegancsmith 2022-09-06
Still lots of fruit to squeeze here, according to CPU profiles we are only spending ~15% of our time actually doing path matching (the BTree stuff below)... I think the next change I will look into is a bit of refactoring here so we don't have so much overhead per file check.

![image](https://user-images.githubusercontent.com/187831/188664740-1dbf9dce-e51d-45f9-8ad8-f9335bf8f289.png)

For others who want to join in the fun quickly, I've been running these commands:

```bash
cd sourcegraph/internal/authz
go test -bench=Filter -run=^$ -cpuprofile cpu.ppro
go tool pprof -call_tree -http=:8081 authz.test cpu.pprof
```