@kopancek 2022-09-06
Should we make caching the new default to create prometheus metrics? At least for counters like this one. E.g. create a wrapper function around it? Because I think this is going to bite us in the future in other sensitive parts of code...