@sashaostrikov 2022-09-06
@kopancek Sounds like a great idea