@keegancsmith 2022-09-07
> Should we make caching the new default to create prometheus metrics? At least for counters like this one. E.g. create a wrapper function around it? Because I think this is going to bite us in the future in other sensitive parts of code...

TBH if you are measuring something where the overhead of measurements is bigger than what you are measuring... it isn't actually worth measuring. So I'd honestly opt to drop this metric instead. However, I think after the refactors I mentioned earlier in this PR, then there will be useful metrics to track where we only increase it a constant times per request. That would be more useful.