@sashaostrikov 2022-09-07
@keegancsmith `time.Now` definitely takes a significant runtime of `Permissions` function, dropping it (and therefore dropping the metric) is a nice speed-up. I don't remember using this metric very frequently, what would @ryanslade and @mollylogue say about it. Is it worth dropping for the sake of performance?