@ryanslade 2022-09-07
I think it makes sense to move the instrumentation higher up the stack. I don't think we need to be instrumenting in the `Permissions` method since it is often called in tight loops.